+++ internal/services/service.go
@@ -36,5 +36,5 @@ type UsersData interface {
 type Workouts interface {
 	Generate(models.User) error
 	GetWorkout(id int) (models.Workout, error)
-	InsertWorkout(workout models.Workout) (int, error)
+	InsertWorkout(workout models.Workout) (int64, error)
 }
